changed height, weight and age properties from double to int

+++ Fitbook/Models/FitbookUser.cs
@@ -13,9 +13,9 @@ namespace Fitbook.Models
         public int FitbookUserId { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
-        public double? Age { get; set; }
-        public double? Height { get; set; }
-        public double? Weight { get; set; }
+        public int? Age { get; set; }
+        public int? Height { get; set; }
+        public int? Weight { get; set; }
         public string Gender { get; set; }
         public string Lifestyle { get; set; }
         public string ApplicationUserId { get; set; }
